fix: vector-based FieldOffsets
(RawBufferExpandsIntoMultipleBuffer.test) now passing
Prior, we had to use multiple tuple buffers in case not all indexes fit into a single buffer.
This added significant complexity in writing to and reading from the index buffers.
Now, we replaced the uses of TupleBuffers in FieldOffsets with a vector. Since the vector dynamically resizes, we can read from a single address.